requirements.txt
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
rich>=13.0.0
inquirer>=3.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
apscheduler>=3.10.0
pyyaml>=6.0
jinja2>=3.1.0
pytest>=7.4.0
pytest-mock>=3.12.0

